clean up performancev2 and pipeline funcitons
