+++ backend/data/updatepipeline.py
@@ -4,9 +4,6 @@ import time
 import performancev2
 import pandas as pd
 import datetime as dt
-from functools import partial
-from bson import ObjectId
-from billiard.pool import Pool
 
 TIME_ZONE_OFFSET = -dt.timedelta(hours=7)
 
@@ -106,11 +103,9 @@ def proximity_update(update_type, batch_size=100, wait=True, additional_query=No
             if update_type == 'volume':
                 if place['activity'] == []:
                     continue
-                local_retail_volume = compile_details(place['location'],
-                                                      performancev2.LOCAL_RETAIL_RADIUS)
-                local_category_volume = compile_details(
-                    place['location'], performancev2.LOCAL_CATEGORY_RADIUS,
-                    retail_type=place['type']) if 'type' in place else -1
+                local_retail_volume = performancev2.local_retail_volume(place['location'])
+                local_category_volume = performancev2.local_category_volume(
+                    place['location'], place['type']) if 'type' in place else -1
 
                 results.append({
                     '_id': place['_id'],
@@ -118,7 +113,7 @@ def proximity_update(update_type, batch_size=100, wait=True, additional_query=No
                     'local_category_volume': local_category_volume
                 })
             elif update_type == 'confidence':
-                num_nearby = len(get_nearby(place['location'], 0.01))
+                num_nearby = len(performancev2.get_nearby(place['location'], 0.01))
                 results.append({
                     '_id': place['_id'],
                     'num_nearby': num_nearby
@@ -144,38 +139,6 @@ def proximity_update(update_type, batch_size=100, wait=True, additional_query=No
         })
 
 
-def compile_details(geo_point, radius, retail_type=None, terminal_db=None):
-
-    nearby_places = get_nearby(geo_point, radius, retail_type=retail_type, terminal_db=terminal_db)
-    volume_array = [place['activity_volume'] for place in nearby_places
-                    if 'activity_volume' in place and place['activity_volume'] > 0]
-    total_volume = sum(volume_array) / len(volume_array) if volume_array else -1
-
-    return total_volume
-
-
-def get_nearby(geo_point, radius, retail_type=None, terminal_db=None):
-
-    if not terminal_db:
-        terminal_db = utils.DB_TERMINAL_PLACES
-
-    query = {}
-    if retail_type:
-        query['type'] = retail_type
-    query.update({
-        'location': {
-            '$near': {
-                '$geometry': geo_point,
-                '$maxDistance': utils.miles_to_meters(radius)
-            }
-        },
-    })
-
-    nearby_places = list(terminal_db.find(query))
-
-    return nearby_places
-
-
 def ordered_update():
 
     for name in pd.read_csv('scripts/files/activity_generated/sorted_names.csv').set_index('_id').index[:1000]:
@@ -197,25 +160,5 @@ def merge_update():
 
 
 if __name__ == "__main__":
-
-    def test_compile_details(burner=None):
-        place = utils.DB_TERMINAL_PLACES.find_one({'_id': ObjectId("5eca2c36eabaf79dfe0825f1")})
-        total_volume = compile_details(place['location'], 1)
-        total_volume_category = compile_details(place['location'], 3, place['type'])
-        assert abs(total_volume - place['local_retail_volume']) < .5
-        assert abs(total_volume_category - place['local_category_volume']) < .5
-        print('Success')
-        return(1)
-
-    def test_concurrency():
-        first = time.time()
-        pool = Pool(40)
-        pool.map(test_compile_details, list(range(1000)))
-
-        last = time.time()
-        print(last - first)
-        pool.join()
-        pool.close()
-
     # setup_confidence()
     proximity_update('confidence', wait=False)
